Split out test code from library code
