+++ test/tsconfig.test.json
@@ -11,5 +11,6 @@
   "include": [
     "../src/**/*",
     "../test/helpers/*",
+    "../test/unit/*"
   ]
 }
